js out of appjs
js for mongoose is out of appjs file
